+++ CHANGES.md
@@ -2,11 +2,17 @@
 
 ## unreleased
 
+- 
+
+
 ## 1.0.0
 
 - Distributed authorization/authentication support.
+- Show activity information when loading a new instance.
+
 
 ## 0.113 (?)
+
 - Log in on EduVPN
 - Handoff OpenVPN certificate to OpenVPN Connect app.
 - Updated look and feel with big headers.